large structural changes to put useContext in MainContentSection and update workorders on state change
