+++ src/components/root/MainContentSection.js
@@ -1,5 +1,6 @@
-import React, { useContext, Fragment } from "react";
+import React, { useContext, Fragment, useCallback } from "react";
 import { UserContext } from "../../backend/authorization/UserContext";
+import { ImpersonatorContext } from "../../backend/authorization/ImpersonatorContext";
 import { Redirect, Route, Switch } from "react-router-dom";
 import Home from "./Home";
 import SidebarNav from "./SidebarNav";
@@ -10,7 +11,37 @@ import AdvertHero from "./AdvertHero";
 import AuthContainer from "../../backend/authorization/AuthContainer";
 
 function MainContentSection({ loginClick }) {
-  const [currentProfile] = useContext(UserContext);
+  const [admin, setAdmin] = useContext(ImpersonatorContext);
+  const [currentProfile, setCurrentProfile] = useContext(UserContext);
+
+
+
+  const loadUserProfile = useCallback(
+    async function () {
+      try {
+        const response = await fetch("/api/profile/this-profile", {
+          headers: {
+            credentials: "include",
+          },
+        });
+        const json = await response.json();
+        if (!response.ok) {
+          throw new Error(json.message);
+        }
+        setCurrentProfile(json.data);
+      } catch (err) {
+        console.log(err);
+        setCurrentProfile(undefined);
+      }
+    },
+    [setCurrentProfile]
+  );
+
+  // function loadUserProfile() {
+
+  //   const prevWorkorders = currentProfile.workorders;
+  //   console.log(`prev: ${prevWorkorders} & new: ${workorders}`);
+  // }
 
   let content;
   if (loginClick) {
@@ -26,8 +57,8 @@ function MainContentSection({ loginClick }) {
               {/* prettier-ignore */}
               <Switch>
                 <Route path="/home" render={() => <Home />} />
-                <Route path="/profile" render={() => <Profile currentProfile={currentProfile} />} />
-                <Route path="/workorder" render={() => <Workorder />} />
+                <Route path="/profile" render={() => <Profile admin={admin} currentProfile={currentProfile} />} />
+                <Route path="/workorder" render={() => <Workorder currentProfile={currentProfile} loadUserProfile={loadUserProfile} />} />
                 <Route path="/settings" render={() => <Settings />} />
               </Switch>
             </div>
@@ -35,7 +66,7 @@ function MainContentSection({ loginClick }) {
         </div>
       );
     } else {
-      content = <AuthContainer />;
+      content = <AuthContainer loadUserProfile={loadUserProfile} />;
     }
   } else {
     content = (
